feat: impl new fs device
